refactor: 미션7 리뷰반영 - 제품정보컴포넌트 리액트 내 스타일드컴포넌트 제거
